// frontend/src/app/components/clients/components/create-client-modal/create-client.model.ts
import { FormControl } from '@angular/forms';

export interface CreateClientForm {
  name: FormControl<string | null>;
  address: FormControl<string | null>;
  invoiceIsDone: FormControl<boolean | null>;
  showOnMap: FormControl<boolean | null>;
  invoiceAcceptanceDate: FormControl<Date | null | undefined>;
  description: FormControl<string | null>;
  teamId: FormControl<string | null | undefined>;
}